SDK Rework Complete
