+++ WindowsClient/AqHaxCSGO/Forms/EntryForm.resx
