+++ WindowsClient/AqHaxCSGO/Memory/NetvarManager.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Resources;
 using System.Text;
 using System.Threading.Tasks;
+using AqHaxCSGO.Objects.Classes;
 
 namespace AqHaxCSGO.MemoryManagers
 {
@@ -90,53 +91,7 @@ namespace AqHaxCSGO.MemoryManagers
 			this.Offset = offset;
 		}
 	}
-	class CClass
-	{
-		private int this_pointer;
 
-		public CClass SetBase(int this_pointer)
-		{
-			this.this_pointer = this_pointer;
-			return this;
-		}
-
-		public int ClassID
-		{
-			get
-			{
-				return Memory.Read<int>(this_pointer + 0x14);
-			}
-		}
-
-		public string ClassName
-		{
-			get
-			{
-				return Memory.ReadString(Memory.Read<int>(this_pointer + 0x8), 64, Encoding.UTF8);
-			}
-		}
-
-		public int Next
-		{
-			get
-			{
-				return Memory.Read<int>(this_pointer + 0x10);
-			}
-		}
-
-		public int Table
-		{
-			get
-			{
-				return Memory.Read<int>(this_pointer + 0xC);
-			}
-		}
-
-		public bool Readable()
-		{
-			return Memory.CanRead(this_pointer, 0x28);
-		}
-	}
 	class RecvProp
 	{
 		private int this_pointer;
