+++ WindowsClient/AqHaxCSGO/Misc/SaveManager.cs
