+++ WindowsClient/AqHaxCSGO/Misc/Threads.cs
