+++ WindowsClient/AqHaxCSGO/Objects/Managers/GlowObjectManager.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace AqHaxCSGO.Objects.Managers
+{
+    static class GlowObjectManager
+    {
+        //Implement Later
+    }
+}
