--- WindowsClient/AqHaxCSGO/Objects/Structs/Hitbox.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace AqHaxCSGO.Objects.Structs
-{
-    /**
-     * All bones will be implemented later on !
-     */
-    public enum Hitbox
-    {
-        ENTITY_HEAD = 8,
-        ENTITY_BODY = 0,
-        ENTITY_RIGHTSHOULDER,
-        ENTITY_LEFTSHOULDER
-    }
-}
