+++ WindowsClient/AqHaxCSGO/Resources/clientIcon.ico
